+++ pyemma/plots/tests/test_plots1d.py
@@ -19,24 +19,36 @@
 from __future__ import absolute_import
 import unittest
 import numpy as np
-import warnings
 
 from pyemma.plots.plots1d import plot_feature_histograms
+import matplotlib.pyplot as plt
 
 
-class TestPlots2d(unittest.TestCase):
+class TestPlots1d(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
         cls.data = np.random.rand(500, 10)
 
     def test_feature_histograms(self):
-        plot_feature_histograms(self.data)
-
-    def test_feature_histograms_nowarning(self):
-        with warnings.catch_warnings(record=True) as w:
-            plot_feature_histograms(self.data)
-            assert len(w) == 0
+        fig, _ = plot_feature_histograms(self.data)
+        plt.close(fig)
 
     def test_invalid_input(self):
         with self.assertRaises(ValueError):
-            plot_feature_histograms(self.data, feature_labels=np.random.rand(5))
+            fig, _ = plot_feature_histograms(self.data, feature_labels=np.random.rand(5))
+            plt.close(fig)
+
+    def test_feature_histograms_mpl_arguments(self):
+        labels = ['PyEMMA' for _ in range(self.data.shape[1])]
+        fig, _ = plot_feature_histograms(self.data,
+                                         feature_labels=labels,
+                                         ylog=True,
+                                         n_bins=10,
+                                         color='g')
+        plt.close(fig)
+
+    def test_feature_histograms_ax_argument(self):
+        from matplotlib.pyplot import subplots
+        fig, ax = subplots()
+        fig, _ = plot_feature_histograms(self.data, ax=ax)
+        plt.close(fig)